+++ src/components/read/TopBar.js
@@ -1,14 +1,14 @@
 import React from "react";
-import AppBar from "@material-ui/core/AppBar";
-// import Button from "@material-ui/core/Button";
-import Toolbar from "@material-ui/core/Toolbar";
-// import InputBase from "@material-ui/core/InputBase";
+import { Link } from "react-router-dom";
 import { fade, makeStyles } from "@material-ui/core/styles";
-// import SearchIcon from "@material-ui/icons/Search";
+import AppBar from "@material-ui/core/AppBar";
+import Toolbar from "@material-ui/core/Toolbar";
 import Typography from "@material-ui/core/Typography";
 import IconButton from "@material-ui/core/IconButton";
-import { Link } from "react-router-dom";
 import MenuIcon from "@material-ui/icons/Menu";
+// import Button from "@material-ui/core/Button";
+// import InputBase from "@material-ui/core/InputBase";
+// import SearchIcon from "@material-ui/icons/Search";
 
 const useStyles = makeStyles(theme => ({
   root: {
